refactoring migrations

+++ books-maslov/src/main/resources/schema/1.0.0/tables/book.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS book (
-    id int NOT NULL UNIQUE,
+    id bigserial NOT NULL UNIQUE,
     name text,
     author_id text,
     year_of_publishing text,
