+++ books-maslov/src/main/resources/schema/1.0.0/tables/genre.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS genre (
-    id int,
+    id bigserial,
     name text,
     PRIMARY KEY (id)
 )
